Assignment 5
Book Chapter 6 Question 2 and 4. Classes and constructors.
